move getTime to utils / remove br's
